Rename mlflow_faculty.artifact_repository -> mlflow_faculty.artifacts
